+++ Engine/PersistentPedalOverridePatch.cpp
@@ -1,6 +1,6 @@
 /*
  * mTroll MIDI Controller
- * Copyright (C) 2015 Sean Echevarria
+ * Copyright (C) 2015,2025 Sean Echevarria
  *
  * This file is part of mTroll.
  *
@@ -25,53 +25,92 @@
 #include "PersistentPedalOverridePatch.h"
 
 
-PersistentPedalOverridePatch* PersistentPedalOverridePatch::sActiveOverride = nullptr;
+PersistentPedalOverridePatch* PersistentPedalOverridePatch::sActiveOverride[ExpressionPedals::PedalCount] = { nullptr };
 ExpressionPedals* PersistentPedalOverridePatch::sInactivePedals = nullptr;
+ExpressionPedalAggregate* PersistentPedalOverridePatch::sAggregateOverridePedals = nullptr;
 
 
 void
 PersistentPedalOverridePatch::SwitchPressed(IMainDisplay * mainDisplay, ISwitchDisplay * switchDisplay)
 {
-	PersistentPedalOverridePatch * prev = sActiveOverride;
+	const PersistentPedalOverridePatch * prev[ExpressionPedals::PedalCount] = { sActiveOverride[0], sActiveOverride[1], sActiveOverride[2], sActiveOverride[3] };
 
 	__super::SwitchPressed(mainDisplay, switchDisplay);
 
-	if (prev && prev != sActiveOverride)
+	for (int idx = 0; idx < ExpressionPedals::PedalCount; ++idx)
 	{
-		// update switch of override patch that this one replaced
-		prev->UpdateDisplays(nullptr, switchDisplay);
+		if (prev[idx] && prev[idx] != sActiveOverride[idx])
+		{
+			// update switch of override patch that this one replaced
+			prev[idx]->UpdateDisplays(nullptr, switchDisplay);
+		}
 	}
 }
 
 void
 PersistentPedalOverridePatch::ExecCommandsA()
 {
-	if (sActiveOverride && sActiveOverride != this)
+	for (int idx = 0; idx < ExpressionPedals::PedalCount; ++idx)
 	{
-		// update state of override patch that this one replaced
-		sActiveOverride->ExecCommandsB();
-		sInactivePedals = gActivePatchPedals;
-		gActivePatchPedals = &mPedals;
+		if (sActiveOverride[idx] && sActiveOverride[idx] != this && 
+			(mPedals.IsLocalPedalEnabled(idx) || mPedals.IsGlobalPedalDisabled(idx)))
+		{
+			auto prev = sActiveOverride[idx];
+
+			// replace the old override -- to short-circuit the reset logic in ExecCommandsB
+			sActiveOverride[idx] = this;
+
+			// update state of override patch that this one replaced
+			prev->OverridePedals(true);
+			prev->ExecCommandsB();
+			prev->OverridePedals(false);
+		}
 	}
-	else if (gActivePatchPedals != &mPedals)
+
+	if (sAggregateOverridePedals)
+		sAggregateOverridePedals->OverridePedals(mPedals);
+
+	if (gActivePatchPedals != sAggregateOverridePedals)
+	{
 		sInactivePedals = gActivePatchPedals;
+		gActivePatchPedals = sAggregateOverridePedals;
+	}
 
+	OverridePedals(true);
 	__super::ExecCommandsA();
+	OverridePedals(false);
 
-	_ASSERTE(gActivePatchPedals == &mPedals);
-	sActiveOverride = this;
+	_ASSERTE(gActivePatchPedals == sAggregateOverridePedals); // assert if the exec changed it underneath us
+	for (int idx = 0; idx < ExpressionPedals::PedalCount; ++idx)
+		if ((mPedals.IsLocalPedalEnabled(idx) || mPedals.IsGlobalPedalDisabled(idx)) && sActiveOverride[idx] != this)
+			sActiveOverride[idx] = this;
 }
 
 void
 PersistentPedalOverridePatch::ExecCommandsB()
 {
-	_ASSERTE(sActiveOverride == this);
-	_ASSERTE(gActivePatchPedals == &mPedals);
+	_ASSERTE(gActivePatchPedals == sAggregateOverridePedals);
 
+	OverridePedals(true);
 	__super::ExecCommandsB();
+	OverridePedals(false);
 
 	_ASSERTE(sInactivePedals != &mPedals);
+	_ASSERTE(sInactivePedals != sAggregateOverridePedals);
+
+	if (sAggregateOverridePedals)
+		sAggregateOverridePedals->ClearPedalOverrides(mPedals);
+
+	for (int idx = 0; idx < ExpressionPedals::PedalCount; ++idx)
+		if (sActiveOverride[idx] == this && (mPedals.IsLocalPedalEnabled(idx) || mPedals.IsGlobalPedalDisabled(idx)))
+			sActiveOverride[idx] = nullptr;
+
+	// see if any other overrides are active, return if so
+	for (auto & cur : sActiveOverride)
+		if (cur)
+			return;
+
+	// else, reset to default non-persistent behavior
 	gActivePatchPedals = sInactivePedals;
 	sInactivePedals = nullptr;
-	sActiveOverride = nullptr;
 }